+++ packages/test/src/utilities/textualizer.test.ts
@@ -1,4 +1,5 @@
-import { serializerWith } from "@lionweb/core"
+import { Classifier, LanguageFactory, serializerWith } from "@lionweb/core"
+import { concatenator } from "@lionweb/ts-utils"
 import { genericAsTreeText, languageAsText } from "@lionweb/utilities"
 import { readFileSync, writeFileSync } from "fs"
 
@@ -14,6 +15,31 @@ describe("LionCore-specific textual syntax", () => {
         const thisPath = "src/utilities"
         writeFileSync(`${thisPath}/languageWithEnum.actual.txt`, actual)
         equal(actual, readFileSync(`${thisPath}/languageWithEnum.expected.txt`, { encoding: "utf8" }).replaceAll("\r\n", "\n"))
+    })
+
+    it("textualize a language with unset references", () => {
+        const factory = new LanguageFactory("foo", "0", concatenator("-"), concatenator("-"))
+        factory.annotation("Anno")   // no .annotates
+        const concept = factory.concept("Bar", false)
+        factory.property(concept, "tender") // no .type
+        factory.reference(concept, "lizzard")   // no .type
+        equal(
+            languageAsText(factory.language),
+            `language foo
+    version: 0
+    entities (↓name):
+
+        annotation Anno
+            annotates: ???
+
+        concept Bar
+            features (↓name):
+                lizzard -> ???
+                tender: ???
+
+
+`
+        )
     })
 })
 
@@ -47,6 +73,32 @@ describe("generic textual syntax", () => {
     books:
         Book (id: Q1NQ_zXwOF2Pjf1uklMTumJhQH4V1Qnfjnu0gz7Fzzs) {
             title = 'Explorer Book'
+            pages = 1337
+            type = Special
+            author -> 7RxOqvS1ZIdpk-ao-6Tzy1QQRxl6fp_tGMH_BIK4LzQ (Jack London)
+        }
+}
+GuideBookWriter (id: 7RxOqvS1ZIdpk-ao-6Tzy1QQRxl6fp_tGMH_BIK4LzQ) {
+    countries = 'Alaska'
+    name = 'Jack London'
+}
+`
+        )
+    })
+
+    it("textualize library model with broken (incomplete) language def.", () => {
+        (libraryLanguage.entities
+            .filter((entity) => entity instanceof Classifier)
+            .find((entity) => entity.features.length > 0)!
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            .features[0] as any)["type"] = undefined
+        equal(
+            genericAsTreeText(serializerWith({ reader: libraryReader })(libraryModel), [libraryLanguage]),
+            `Library (id: jkxERSov0TuSh7MMz2D5ciLrZDrU4o_VMOpmBqh_j7E) {
+    name = 'Bob's Library'
+    books:
+        Book (id: Q1NQ_zXwOF2Pjf1uklMTumJhQH4V1Qnfjnu0gz7Fzzs) {
+            title = ???
             pages = 1337
             type = Special
             author -> 7RxOqvS1ZIdpk-ao-6Tzy1QQRxl6fp_tGMH_BIK4LzQ (Jack London)