+++ packages/utilities/src/m3/textualizer.ts
@@ -24,10 +24,10 @@ import { asString, indentWith, Template } from "littoral-templates"
 const indented = indentWith("    ")(1)
 
 const refAsText = <NT extends INamed & Node>(ref: SingleRef<NT>): string =>
-    isUnresolvedReference(ref) ? `???` : ref.name
+    (ref === undefined || isUnresolvedReference(ref)) ? `???` : ref.name
 
 const recurse = <NT extends M3Node>(ts: NT[], header: string, func: (t: NT) => Template = asText): Template =>
-    ts.length === 0 ? [] : indented([header, indented(ts.map(func))])
+    (ts === undefined || ts.length === 0) ? [] : indented([header, indented(ts.map(func))])
 
 const featuresOf = (classifier: Classifier): Template => recurse(nameSorted(classifier.features), `features (↓name):`)