+++ packages/core/CHANGELOG.md
@@ -10,6 +10,8 @@
     * Introduce `Deserializer` and `DeserializerConfiguration` types, and a `deserializerWith` function as a replacement (keeping an alias for backward compatibility) for the `deserializeSerializationChunk` function with a better API (using the Parameter Object pattern).
     * Rename `SimplisticHandler` interface to `ProblemReporter`, and propagate that renaming, keeping alias for backward compatibility.
 * Introduce an explicit `UnresolvedReference` type, a corresponding `isUnresolvedReference` type function, and a `referenceToSet` function to produce reference (target) placeholders, and propagate their use throughout the codebase.
+* Add a 2nd argument to the `ResolveInfoDeducer` function type, to also pass the containing reference.
+    *Note* that this is technically a breaking change, but the `ResolveInfoDeducer` type is essentially part of the internal API.
 
 
 ## 0.7.2